Added support for specifying multiple PK's on a poco.     [PrimaryKey("pkField1", "pkField2", ...)]
Changes PrimaryKey on TableInfo to an array.  Referencing code will need to be changed to take the array and merge the key's to create a multi-part conditional instead of just using it directly.  IE:
   WHERE pkField1 = @1 AND pkField2 = @2

Updated user accessible functions to take PrimaryKey values as a list or params.
   Database.Exists<Table>(pkValue1, pkValue2, ...)
     .Single, .SingOrDefault, .Update, .Delete
  Note that the values in these cases need to be order in the same order that the Primary Key fields are specified.

  This enables code like this for the user:
    Database.delete(1, 3);

Still needs to be updated:
  Internal code and functions that reference and apply filters based off the Primary Key fields.
  Tests and validation.
  T4 Templates to detect and include multiple key's when detected.