Reinstate virtual-ness of OnExecutedCommand
The `virtual` keyword was removed in `7de9a0e13a41` "Completes BeginTransactionAsync and PageAsync", as part of PR #507 "Add async support".
I can't see any reason in the surrounding commit to have removed access to this hook-in point, and the other half of the pair (`OnExecutingCommand`) is still virtual.

Perhaps is was simply unintentional? but if intentional then my only guess is that @pleb thought that it wasn't a useful hook point - that "before the command runs" was the useful time to do things :)
I'll respectfully disagree - this pair of hook points allows for an easy place to put things "around" every command, e.g. Perf tracing code and other NFR stuff. I agree that there's very little that you'd want to do to *modify* the cmd object, but having it as essentially an event hook, is still valuable.